Move refreshSourceSchemaIfNeeded from Crud to DdlManager
